word/requirements.txt
openai-whisper
sounddevice
numpy
scipy
soundfile
